Add support for i386
CPU architecture can now be denotated as the first parameter of
`rootfs-builder`. If none is provided, it defaults to the host's
architecture. For now, only amd64 and i386 will be supported.
These are the only two architectures that I will ever use with Debian
and I don't want to support anything else. For Raspbian, it would make
sense to make a separate script (and corresponding packer builds). If I
ever start using SBC's other than Raspberry Pi's, I will reconsider
supporting armhf for this script.